moved comp to app
