--- client/src/components/Header.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const Header = () => {
-  return (  <div className="peachHeader">
-    <div className="peachSide">
-      <h1 className="peachFont peachTitle">Son of a Peach</h1>
-    </div>
-    <div className="peachSide">
-    </div>
-    <h5 className="peachFont peachSubtitle">click the peach to free him!</h5>
-  </div>
-  );
-};
-
-export default Header;
